+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+
+## 4.2.0 - 2022-10-24
 - Allow to use custom progress bar instance
     - Add `IProgress` interface for a progress bar
     - Add `ConsoleStyle` method